requirements.txt
Django==4.1.3
celery==5.2.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
channels==4.0.0
channels-redis==4.0.0
django-environ==0.7.0
django-timezone-field==5.0
flower==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
gunicorn==20.1.0
humanize==4.4.0
kombu==5.2.4
msgpack==1.0.4
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
django-import-export==3.0.1